Addition of the grid squares
The grid squares are created using a for
loop, which generates all the divs and adds
them to the div with the class name 'content'.
In the CSS, there are class selectors
'content' that define the style of the main
div and 'grid' that define the style of the
divs generated by JavaScript.