Refactor of the form which creates the divs
A method called squareDraw that has the form modified within and
has two parameters: a width and a height. The changes in the form
are: the limit of the loop are the multiplication of the
parameters and the addition of the width and height in the div's
styles whose values are the total width or heigth of the parent
div divided by the width or height.